ecommerce/requirements.txt
Django==3.1.4
gunicorn==20.0.4
django-import-export==2.4.0
django-filter==2.4.0
mysqlclient==2.0.2
Pillow==8.0.1
